+++ src/tools/index.js
@@ -1,26 +1,24 @@
 import delay from './delay';
 import genNonDuplicateID from './genNonDuplicateID';
 import getUrlParameter from './getUrlParameter';
-import getUrlParams from './getUrlParams';
 import momentToString from './momentToString';
 import toThousands from './toThousands';
 import objectToString from './objectToString';
 import noPassByInfo from './noPassByInfo';
 import countdown from './countdown';
 import storageWorker from './storageWorker';
-import isPromise from './isPromise';
 import useDebounce from './useDebounce';
 import clipboard from './clipboard';
+import typeOf from './typeOf';
 
 export default {
   delay,
-  isPromise,
+  typeOf,
   countdown,
   clipboard,
   toThousands,
   useDebounce,
   noPassByInfo,
-  getUrlParams,
   storageWorker,
   momentToString,
   objectToString,