+++ src/tools/toThousands.js
@@ -1,5 +1,3 @@
-import { warning } from '../_util';
-
 function toThousands(v = 0, figures) {
   if (typeof +v === 'number' && !isNaN(v) ) {
     const arr = `${v}`.split('.'); // 将数字进行整数位和小数位的分隔
@@ -11,7 +9,7 @@ function toThousands(v = 0, figures) {
     try {
       return +v.replace(/\$\s?|(,*)/g, '');
     } catch (e) {
-      warning('toThousands', 'Parameter type error');
+      console.error(`Warning: [toThousands] Parameter type error`);
       return false;
     }
   }