+++ docs/changeLog.md
@@ -6,6 +6,17 @@ title: 更新日志
 
 遵循 [Semantic Versioning 2.0.0](https://semver.org/lang/zh-CN/) 语义化版本规范。
 
+### 🚀 2.9.0
+
+`2021-05-13`
+
+- tools
+  - 删除 getUrlParams 方法，现在由 getUrlParameter 替代
+  - 删除 isPromise 方法，现在由 typeOf 替代
+  - 优化 noPassByInfo，现在支持 "身份证、姓名、手机号、邮箱" 的脱敏了
+- 移除了 Role 组件，请在项目中自行消化
+- 优化 DateEasily 的文档
+
 ### 🚀 2.8.12
 
 `2021-05-10`